requirements.txt
app-observer==0.1
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
